pintrail/worker/pyproject.toml
[project]
name = "wk2"
version = "0.1.0"
description = "Pintrail image-processing worker"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "arq>=0.26.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.11",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
